Successful rejection detected in tx list
